refactor: extract platform path detection to utility, move autodetect into LocationsTabController
Extract _find_steam_root, __get_darwin_paths, __get_linux_paths, and
__get_windows_paths from SettingsController into a pure utility module
(app/utils/steam_path_detection.py) with no Qt dependencies. The
autodetect button handler now lives directly in LocationsTabController,
removing the callback indirection. Tests migrated from name-mangled
SettingsController method calls to direct utility function calls.
