+++ tests/controllers/test_path_autodetection.py
@@ -1,7 +1,11 @@
 from pathlib import Path
 from unittest.mock import patch
 
-from app.controllers.settings_controller import SettingsController
+from app.utils.steam_path_detection import (
+    find_steam_root,
+    get_darwin_paths,
+    get_linux_paths,
+)
 
 
 def _vdf_escape(path: Path) -> str:
@@ -18,24 +22,15 @@ def _setup_steam_root(tmp_path: Path, subdir: str) -> Path:
     return steam_root
 
 
-def _make_controller() -> SettingsController:
-    """Create a SettingsController without __init__ (which requires Qt widgets).
-
-    Safe because the path detection methods only use Path.home(), static
-    methods, and module-level utility functions — no instance attributes.
-    """
-    return SettingsController.__new__(SettingsController)
-
-
 class TestFindSteamRoot:
-    """Tests for SettingsController._find_steam_root()."""
+    """Tests for find_steam_root()."""
 
     def test_returns_first_valid_candidate_with_steamapps(self, tmp_path: Path) -> None:
         candidate = tmp_path / ".steam" / "steam"
         candidate.mkdir(parents=True)
         (candidate / "steamapps").mkdir()
 
-        result = SettingsController._find_steam_root([candidate])
+        result = find_steam_root([candidate])
         assert result == candidate
 
     def test_returns_first_valid_candidate_with_vdf(self, tmp_path: Path) -> None:
@@ -43,7 +38,7 @@ class TestFindSteamRoot:
         (candidate / "config").mkdir(parents=True)
         (candidate / "config" / "libraryfolders.vdf").touch()
 
-        result = SettingsController._find_steam_root([candidate])
+        result = find_steam_root([candidate])
         assert result == candidate
 
     def test_skips_nonexistent_candidates(self, tmp_path: Path) -> None:
@@ -52,18 +47,18 @@ class TestFindSteamRoot:
         valid.mkdir()
         (valid / "steamapps").mkdir()
 
-        result = SettingsController._find_steam_root([nonexistent, valid])
+        result = find_steam_root([nonexistent, valid])
         assert result == valid
 
     def test_skips_candidates_without_steamapps_or_vdf(self, tmp_path: Path) -> None:
         empty_dir = tmp_path / "empty"
         empty_dir.mkdir()
 
-        result = SettingsController._find_steam_root([empty_dir])
+        result = find_steam_root([empty_dir])
         assert result is None
 
     def test_returns_none_when_no_candidates_match(self, tmp_path: Path) -> None:
-        result = SettingsController._find_steam_root([tmp_path / "a", tmp_path / "b"])
+        result = find_steam_root([tmp_path / "a", tmp_path / "b"])
         assert result is None
 
     def test_respects_priority_order(self, tmp_path: Path) -> None:
@@ -75,44 +70,44 @@ class TestFindSteamRoot:
         second.mkdir()
         (second / "steamapps").mkdir()
 
-        result = SettingsController._find_steam_root([first, second])
+        result = find_steam_root([first, second])
         assert result == first
 
     def test_returns_none_for_empty_list(self) -> None:
-        result = SettingsController._find_steam_root([])
+        result = find_steam_root([])
         assert result is None
 
 
 class TestGetLinuxPaths:
-    """Tests for SettingsController.__get_linux_paths() via the name-mangled accessor."""
-
-    def _call(self, controller: SettingsController) -> tuple[Path, Path, Path]:
-        return controller._SettingsController__get_linux_paths()  # type: ignore[attr-defined]
+    """Tests for get_linux_paths()."""
 
     def test_debian_installation_path(self, tmp_path: Path) -> None:
         steam_root = _setup_steam_root(tmp_path, ".steam/debian-installation")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
-        assert result[2] == steam_root / "steamapps" / "workshop" / "content" / "294100"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
+        assert (
+            result.steam_mods_folder
+            == steam_root / "steamapps" / "workshop" / "content" / "294100"
+        )
 
     def test_native_steam_path(self, tmp_path: Path) -> None:
         steam_root = _setup_steam_root(tmp_path, ".steam/steam")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
 
     def test_local_share_path(self, tmp_path: Path) -> None:
         steam_root = _setup_steam_root(tmp_path, ".local/share/Steam")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
 
     def test_flatpak_path(self, tmp_path: Path) -> None:
         steam_root = _setup_steam_root(
@@ -121,17 +116,17 @@ class TestGetLinuxPaths:
         )
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
 
     def test_snap_path(self, tmp_path: Path) -> None:
         steam_root = _setup_steam_root(tmp_path, "snap/steam/common/.local/share/Steam")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
 
     def test_vdf_finds_rimworld_in_secondary_library(self, tmp_path: Path) -> None:
         steam_root = tmp_path / ".steam" / "steam"
@@ -150,11 +145,12 @@ class TestGetLinuxPaths:
         (vdf_dir / "libraryfolders.vdf").write_text(vdf_content)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == secondary_lib / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == secondary_lib / "steamapps" / "common" / "RimWorld"
         assert (
-            result[2] == secondary_lib / "steamapps" / "workshop" / "content" / "294100"
+            result.steam_mods_folder
+            == secondary_lib / "steamapps" / "workshop" / "content" / "294100"
         )
 
     def test_proton_config_takes_priority(self, tmp_path: Path) -> None:
@@ -177,9 +173,9 @@ class TestGetLinuxPaths:
         proton_config.mkdir(parents=True)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[1] == proton_config
+        assert result.config_folder == proton_config
 
     def test_native_config_when_no_proton(self, tmp_path: Path) -> None:
         _setup_steam_root(tmp_path, ".steam/steam")
@@ -193,39 +189,36 @@ class TestGetLinuxPaths:
         )
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[1] == native_config
+        assert result.config_folder == native_config
 
     def test_no_steam_root_returns_fallback_paths(self, tmp_path: Path) -> None:
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert "steamapps" in str(result[0])
-        assert "Config" in str(result[1])
-        assert "294100" in str(result[2])
+        assert "steamapps" in str(result.game_folder)
+        assert "Config" in str(result.config_folder)
+        assert "294100" in str(result.steam_mods_folder)
 
     def test_priority_debian_over_native(self, tmp_path: Path) -> None:
         debian_root = _setup_steam_root(tmp_path, ".steam/debian-installation")
         _setup_steam_root(tmp_path, ".steam/steam")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == debian_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == debian_root / "steamapps" / "common" / "RimWorld"
 
 
 class TestGetDarwinPaths:
-    """Tests for SettingsController.__get_darwin_paths().
+    """Tests for get_darwin_paths().
 
     Note: macOS uses "Rimworld" (lowercase w) in hardcoded fallback paths,
     while VDF parsing returns "RimWorld" (capital W). This is fine because
     macOS has a case-insensitive filesystem by default. Tests assert the
     exact casing each code path produces.
     """
-
-    def _call(self, controller: SettingsController) -> tuple[Path, Path, Path]:
-        return controller._SettingsController__get_darwin_paths()  # type: ignore[attr-defined]
 
     @staticmethod
     def _make_darwin_steam_root(tmp_path: Path) -> Path:
@@ -249,11 +242,10 @@ class TestGetDarwinPaths:
         (vdf_dir / "libraryfolders.vdf").write_text(vdf_content)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_darwin_paths()
 
-        # VDF returns "RimWorld" (capital W)
         assert (
-            result[0]
+            result.game_folder
             == steam_root / "steamapps" / "common" / "RimWorld" / "RimworldMac.app"
         )
 
@@ -261,79 +253,71 @@ class TestGetDarwinPaths:
         steam_root = self._make_darwin_steam_root(tmp_path)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_darwin_paths()
 
-        # Fallback uses "Rimworld" (lowercase w) — matches existing behavior
         expected_game = (
             steam_root / "steamapps" / "common" / "Rimworld" / "RimworldMac.app"
         )
-        assert result[0] == expected_game
+        assert result.game_folder == expected_game
 
     def test_config_folder_unchanged(self, tmp_path: Path) -> None:
         self._make_darwin_steam_root(tmp_path)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_darwin_paths()
 
         expected_config = (
             tmp_path / "Library" / "Application Support" / "Rimworld" / "Config"
         )
-        assert result[1] == expected_config
+        assert result.config_folder == expected_config
 
     def test_workshop_folder_derived_from_game(self, tmp_path: Path) -> None:
         self._make_darwin_steam_root(tmp_path)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_darwin_paths()
 
-        assert result[2].parts[-3:] == ("workshop", "content", "294100")
+        assert result.steam_mods_folder.parts[-3:] == ("workshop", "content", "294100")
 
     def test_no_steam_root_returns_hardcoded_paths(self, tmp_path: Path) -> None:
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call(_make_controller())
+            result = get_darwin_paths()
 
-        assert "RimworldMac.app" in str(result[0])
-        assert "Config" in str(result[1])
-        assert "294100" in str(result[2])
+        assert "RimworldMac.app" in str(result.game_folder)
+        assert "Config" in str(result.config_folder)
+        assert "294100" in str(result.steam_mods_folder)
 
 
 class TestSnapWarning:
-    """Tests for Snap detection via _detected_steam_root."""
+    """Tests for Snap detection via steam_root in DetectedPaths."""
 
     def test_snap_steam_root_detected(self, tmp_path: Path) -> None:
         _setup_steam_root(tmp_path, "snap/steam/common/.local/share/Steam")
-        controller = _make_controller()
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            controller._SettingsController__get_linux_paths()  # type: ignore[attr-defined]
+            result = get_linux_paths()
 
-        assert controller._detected_steam_root is not None
-        assert "snap" in controller._detected_steam_root.parts
+        assert result.steam_root is not None
+        assert "snap" in result.steam_root.parts
 
     def test_native_steam_root_not_flagged(self, tmp_path: Path) -> None:
         _setup_steam_root(tmp_path, ".steam/steam")
-        controller = _make_controller()
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            controller._SettingsController__get_linux_paths()  # type: ignore[attr-defined]
+            result = get_linux_paths()
 
-        assert controller._detected_steam_root is not None
-        assert "snap" not in controller._detected_steam_root.parts
+        assert result.steam_root is not None
+        assert "snap" not in result.steam_root.parts
 
     def test_no_steam_root_not_flagged(self, tmp_path: Path) -> None:
-        controller = _make_controller()
-
         with patch("pathlib.Path.home", return_value=tmp_path):
-            controller._SettingsController__get_linux_paths()  # type: ignore[attr-defined]
+            result = get_linux_paths()
 
-        assert controller._detected_steam_root is None
+        assert result.steam_root is None
 
 
 class TestVdfEdgeCases:
     """Tests for VDF parsing edge cases in path autodetection."""
-
-    def _call_linux(self, controller: SettingsController) -> tuple[Path, Path, Path]:
-        return controller._SettingsController__get_linux_paths()  # type: ignore[attr-defined]
 
     def test_malformed_vdf_falls_back_to_default(self, tmp_path: Path) -> None:
         steam_root = tmp_path / ".steam" / "steam"
@@ -344,10 +328,13 @@ class TestVdfEdgeCases:
         (vdf_dir / "libraryfolders.vdf").write_text("this is not valid vdf {{{")
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call_linux(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
-        assert result[2] == steam_root / "steamapps" / "workshop" / "content" / "294100"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
+        assert (
+            result.steam_mods_folder
+            == steam_root / "steamapps" / "workshop" / "content" / "294100"
+        )
 
     def test_vdf_without_rimworld_falls_back(self, tmp_path: Path) -> None:
         steam_root = tmp_path / ".steam" / "steam"
@@ -371,6 +358,6 @@ class TestVdfEdgeCases:
         (vdf_dir / "libraryfolders.vdf").write_text(vdf_content)
 
         with patch("pathlib.Path.home", return_value=tmp_path):
-            result = self._call_linux(_make_controller())
+            result = get_linux_paths()
 
-        assert result[0] == steam_root / "steamapps" / "common" / "RimWorld"
+        assert result.game_folder == steam_root / "steamapps" / "common" / "RimWorld"
